Extra api

+++ src/main/java/com/uber/bookingApp/dto/RideRequestDto.java
@@ -23,7 +23,7 @@ public class RideRequestDto {
     private PointDto dropOffLocation;
     private LocalDateTime requestedTime;
     private Rider rider;
-    private Double fare;
+    private Long fare;
     private PaymentMethod paymentMethod;
     private RideRequestStatus requestStatus;
 }
